very rough xml
This is basically a modified version of EArch.xml with the IO and memory modules as well as the CLBs in roughly the correct locations for X0Y0 of the xilinx basys3 part. I used the EArch definition of a "memory block" as well as the CLBs so the actual definitions that match the Xilinx part still need to be implemented. I was able to run this file by using: 
$VTR_ROOT/vtr_flow/scripts/run_vtr_flow.py     $VTR_ROOT/doc/src/quickstart/blink.v    Xilinx.xml    -temp_dir .     --route_chan_width 100

and 

$VTR_ROOT/vpr/vpr    Xilinx.xml   blink.pre-vpr.blif     --route_chan_width 100 --analysis --disp on